backend/requirements.txt
fastapi
uvicorn[standard]
websockets
google-genai
edge-tts
psutil
python-dotenv
